Move OrderSummary calculations and constants to separate file
